🐋 chore: (FE) - room store와 interview store 분리
