+++ frontend/src/pages/Interviewee/Interviewee.tsx
@@ -8,7 +8,7 @@ import useSafeNavigate from '@hooks/useSafeNavigate';
 import usePreventLeave from '@hooks/usePreventLeave';
 import { webRTCStreamSelector, webRTCUserMapState } from '@store/webRTC.store';
 import { currentVideoTimeState } from '@store/currentVideoTime.store';
-import { docsUUIDState, userRoleSelector } from '@store/room.store';
+import { docsUUIDState, userRoleSelector } from '@store/interview.store';
 
 import { socket } from '@service/socket';
 import mediaStreamer from '@service/mediaStreamer';
